line-chat-bot/chatter/requirements.txt
dataclasses-json
requests
line-bot-sdk
